requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
aniso8601==9.0.1
appdirs==1.4.4
argon2-cffi==21.1.0
astroid==2.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
azure-communication-sms==1.0.1
azure-core==1.19.0
Babel==2.9.1
backcall==0.2.0
Backtesting==0.3.2
backtrader==1.9.76.123
bandit==1.7.0
beautifulsoup4==4.10.0
bitfinex-api-py==1.1.15
bleach==4.1.0
bokeh==2.4.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
coverage==4.5.4
cycler==0.10.0
DateTime==4.3
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
docopt==0.6.2
docutils==0.17.1
dodgy==0.2.1
empyrical==0.5.5
entrypoints==0.3
flake8==3.8.4
flake8-polyfill==1.0.2
Flask==2.0.2
flask-restx==0.5.1
gitdb==4.0.7
GitPython==3.1.24
httplib2==0.19.1
idna==3.2
imagesize==1.2.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==4.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-server==1.0.9
kiwisolver==1.3.2
kiwoom==1.2.1
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.2.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
msrest==0.6.21
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.20.0
oauthlib==3.1.1
packaging==21.0
pandas==0.25.3
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pbr==5.6.0
pep8-naming==0.10.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.18
prospector==1.3.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.1.1
pyee==8.2.2
pyflakes==2.2.0
pyfolio @ git+https://github.com/quantopian/pyfolio.git@4b901f6d73aa02ceb6d04b7d83502e5c6f2e81aa
Pygments==2.9.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
PySimpleGUI==4.45.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==0.5.7
pyxing==0.0.5
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
requirements-detector==0.7
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
setoptconf==0.3.0
simplejson==3.17.5
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stevedore==3.4.0
supervisor==4.2.2
ta==0.7.0
TA-Lib @ file:///C:/aT/alpha-01/TA_Lib-0.4.21-cp38-cp38-win32.whl
terminado==0.9.5
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tulipy==0.4.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.4.0
